+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+- New SideBar module and SideBar class. The sidebar is a qwidget object that
+    contains buttons, metadata and has a background to separate it visually
+    from where it's placed.
+- New SideBarBtn, it is a fixed size solid background button and can
+    contain an icon and a label.
+
+### Removed
+
+### Changed
+
+### Fixed
+
 ## V0.2.0
 ### Added
 - The program now works with sessions that must be started with user data.